requirements.txt
Flask
Flask-SQLAlchemy
Flask-Cors
python-dotenv
requests
pandas
openpyxl